refactor: Move AudioBrowser setup code
- Moved code to `startup` module & break up event & Android Auto setup into different files.
- Verified Android Auto still works.
- As a side-note, we'll remove the "Downsampling Processor" setting in this version (`v3.6.0`), cleaning up the code even further.
